Update integ tests to property simulate clicks
